Final changes, refactored and cleaned code
